+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_08_22_104025) do
+ActiveRecord::Schema[7.0].define(version: 2022_08_23_081934) do
   create_table "actions", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.string "title", null: false
     t.integer "order", null: false
@@ -49,10 +49,10 @@ ActiveRecord::Schema[7.0].define(version: 2022_08_22_104025) do
     t.integer "urn", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.uuid "team_leader_id", null: false
+    t.uuid "team_leader_id"
     t.integer "trust_ukprn", null: false
     t.date "target_completion_date", null: false
-    t.uuid "regional_delivery_officer_id", null: false
+    t.uuid "regional_delivery_officer_id"
     t.uuid "caseworker_id"
     t.index ["caseworker_id"], name: "index_projects_on_caseworker_id"
     t.index ["regional_delivery_officer_id"], name: "index_projects_on_regional_delivery_officer_id"